+++ source/extensions/lightspeed.trex.properties_pane.shared.asset_replacements.widget/lightspeed/trex/properties_pane/shared/asset_replacements/widget/setup_ui.py
@@ -21,10 +21,12 @@ from pathlib import Path
 from typing import Any, Callable, List
 
 import omni.client
+import omni.usd
 from lightspeed.common.constants import GAME_READY_ASSETS_FOLDER as _GAME_READY_ASSETS_FOLDER
 from lightspeed.common.constants import REMIX_CAPTURE_FOLDER as _REMIX_CAPTURE_FOLDER
 from lightspeed.layer_manager.core import LayerManagerCore as _LayerManagerCore
 from lightspeed.layer_manager.core import LayerType as _LayerType
+from lightspeed.trex.material.core.shared import Setup as _MaterialCore
 from lightspeed.trex.material_properties.shared.widget import SetupUI as _MaterialPropertiesWidget
 from lightspeed.trex.mesh_properties.shared.widget import SetupUI as _MeshPropertiesWidget
 from lightspeed.trex.replacement.core.shared import Setup as _AssetReplacementCore
@@ -47,7 +49,7 @@ from omni.flux.utils.common import reset_default_attrs as _reset_default_attrs
 from omni.flux.utils.widget.collapsable_frame import (
     PropertyCollapsableFrameWithInfoPopup as _PropertyCollapsableFrameWithInfoPopup,
 )
-from pxr import Sdf, Tf
+from pxr import Sdf, Tf, Usd
 
 
 class CollapsiblePanels(Enum):
@@ -94,6 +96,7 @@ class AssetReplacementsPane:
         self._context_name = context_name
         self._replacement_core = _AssetReplacementCore(context_name)
         self._layers_core = _AssetReplacementLayersCore(context_name)
+        self._material_core = _MaterialCore(context_name)
 
         self._material_converted_sub = None
 
@@ -289,7 +292,7 @@ class AssetReplacementsPane:
                                 "the weaker layer (bottom).\n",
                                 collapsed=False,
                                 pinnable=True,
-                                pinned_text_fn=lambda: self._get_selection_name_by_type([_ItemPrim]),
+                                pinned_text_fn=lambda: self._get_selection_name_by_type([Usd.Prim]),
                                 unpinned_fn=self._refresh_material_properties_widget,
                             )
                             self._collapsible_frame_states[CollapsiblePanels.MATERIAL_PROPERTIES] = True
@@ -338,19 +341,31 @@ class AssetReplacementsPane:
         return self._selection_tree_widget
 
     def _get_selection_name_by_type(self, selection_types: List):
-        # get a selection based on desired types in descending order from selection_types
-        selection = []
+        context = omni.usd.get_context(self._context_name)
+        stage = context.get_stage()
+        selection_prim_paths = context.get_selection().get_selected_prim_paths()
+
+        # Get all of the material prims from the current selection without duplicates
+        material_prims = set()
+        for prim_path in selection_prim_paths:
+            prim = stage.GetPrimAtPath(prim_path)
+            for material_prim in self._material_core.get_materials_from_prim(prim):
+                material_prims.add(stage.GetPrimAtPath(material_prim))
+        selection = list(material_prims)
+
+        # Get a filtered selection based on desired types in descending order from selection_types
+        type_filtered_selection = []
         for selection_type in selection_types:
-            selection = [
-                item for item in self._selection_tree_widget.get_selection() if isinstance(item, selection_type)
-            ]
-            if selection:
+            type_filtered_selection = [item for item in selection if isinstance(item, selection_type)]
+            if type_filtered_selection:
                 break
 
-        if not selection:
+        if not type_filtered_selection:
             return "None Selected"
-        if len(selection) == 1:
-            formatted_name = "/".join(selection[0].path.split("/")[-2:]).lstrip("/")
+        if len(type_filtered_selection) == 1:
+            # Create parent/child formatted name and limit to 50 chars in length
+            path = type_filtered_selection[0].GetPath()
+            formatted_name = f"{path.GetParentPath().name}/{path.name}"
             return formatted_name if len(formatted_name) < 50 else "..." + formatted_name[-50:]
         return "Multiple Selected"
 
@@ -424,7 +439,11 @@ class AssetReplacementsPane:
     def _refresh_material_properties_widget(self):
         if self._material_properties_collapsable_frame.pinned:
             return
-        items = self._selection_tree_widget.get_selection()
+
+        # Grab the selection prims and refresh the properties
+        context = omni.usd.get_context(self._context_name)
+        prim_paths = list(set(context.get_selection().get_selected_prim_paths()))
+        items = [context.get_stage().GetPrimAtPath(prim_path) for prim_path in prim_paths]
         self._material_properties_widget.refresh(items)
 
     def refresh(self):