+++ source/extensions/lightspeed.trex.material.core.shared/docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0]
+### Changed
+- Added material prim input support for `get_materials_from_prim()`
 
 ## [1.0.3]
 ### Changed