+++ source/extensions/lightspeed.trex.utils.common/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.0]
+## Added
+- Added `get_reference_file_paths` to `prim_utils.py`
+
 ## [1.4.0]
 ## Changed
 - Renamed trex prim utility names for clarity