+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.15.0]
+## Changed
+- Changed `usd_base.py` to use override-able `_get_selection()`
+
 ## [1.14.0]
 ## Added
 - Added a materials interaction tab